requirements.txt
matplotlib
numpy==1.22.4
Pillow==8.4.0
